Adds option to use stylesheet in cards
Still a lot of testing to do
Does allow only one stylesheet yet
Cleanup necessary
