+++ usm/tests/mock-data/usm.renderCards/input-2/usm.json
@@ -1,0 +1,6 @@
+{
+    "activities": [
+        {},
+        {}
+    ]
+}
